fix templatePicker, fix some tslint
